+++ packages/appcd-detect/CHANGELOG.md
@@ -1,3 +1,8 @@
+# v2.2.3 (Jan 10, 2020)
+
+ * fix: Only use a single recursive file watcher instead of watching each found path when recursive
+   and redetect flags are set.
+
 # v2.2.2 (Jan 8, 2020)
 
  * chore: Updated dependencies.